WIP node-forge native implementation in progress
I want to get rid of our external dependency on OpenSSL using
node-forge. But TLS certificates are super complicated.

This generates a readable cert authority and leaf cert, but OSX doesn't
trust it. It says the credentials are invalid. Both the OSX `security`
tool and the `openssl` CLI declare that the client cert isn't trusted.
in OpenSSL's case it says it can't find the issuer.

So I'm

- Unsure if authorityKeyIdentifier is set up correctly
- Unsure if serial number is right
- Still searching for issues in the forge GH repo about trusting root
CAs and signing client CAs with them
- Thinking about adding a Certificate Signing Request back, but Forge's
implementation of that is kind of vague

To test your work, do:

1. Run `yarn build --watch` to start the incremental TS compiler

2. Run `npm link` to symlink this package globally

3. CD to your pwa studio directory

4. Run `npm link devcert` to switch to your local devcert implementation

5. Open Keychain Access

6. Switch to System keychain

7. Find `devcert` and right click to delete. It asks for your password
twice

8. Run `rm -rf ~/Library/Application\ Support/devcert`

9. Clear your SSL cache in browsers (this is in advanced options usually)

10. In PWA Studio directory, run:

  ```sh
  npx lerna --scope=@magento/pwa-buildpack build && npm run watch:venia`
  ```

11. Enter credentials if asked

12. When dev server is ready, visit it in browser

13. If it doesn't trust the cert, it's not fixed. Examine details

14. Make edits to devcert code to fix it

15. Repeat from step 5

You're an angel! Please do this for me and keep me out of the code pit!